Report specifications result for each suite in the right order
Spec is an async method, whether the callback is async or not.
This allows an easy handling for asynchronous callback with async/await.

This ease comes at a cost: when it's time to print results from the
`suite` method, it may happen that some `specs` didn't finish yet.

To solve this we have the following alternatives:
 - use `async` in every suite and `await` in front of every spec;
 - go through the report and every time a pending spec is hit we `await`
   that to finish;
 - use a Promise for each spec and resolve it when spec finishes;
 - create a status for each suite and start printing it only when it's
   not pending;

 I choose the third as it looks more straightforward and does not imply
 any change in how the tests needs to be written.

 STILL MISSING:
 - detailed report for failing specs;
 - support for real test suite (now ony with 'test' one);