Feature/event driven (#17) (#18)
* Create EventBus and moved Spec to Events

* Move Suite to Events

This also involves:
- slight refactor of CLI;
- remove LoggingEventEmitter utility class;
- improve splitting CLI reporter from general report stuff;

* Refactor Event Driven Architecture

Disclaimer: this is a broken commit. I changed my mind one million times
regarding architecture and now I have clarified my mind I need a startin
g point.

This commit might be amended later on.

* Improve separation between core and CLI's setup

Now we have two different `_const` files dealing with its own module.

* Create Process Manager

* Split Database and Reporter

* Fix Database to follow new architecture

* Fix Spec to follow new architecture

* Fix Suite to follow new architecture

* Make EventEmitter be the unique emitter

* Migrate to new architecture, iteration #1

* Migrate to new architecture, iteration #2, working

* Migrate to new architecture, iteration #3

Now EventBus is a utility class meant to be a hub for events. It's not a
n event emitter and requires all other modules, not vice-versa

* Move EventBus to be a function

It was useless to have it as a class...

* Add documentation for new architecture